Split command palette ranker
