+++ Tests/QuillCodeParityTests/ParityGateTests.swift
@@ -1841,11 +1841,18 @@ final class ParityGateTests: XCTestCase {
     func testWorkspaceSurfaceDelegatesCommandPaletteContract() throws {
         let surfaceText = try Self.appSourceText(named: "WorkspaceSurface.swift")
         let paletteText = try Self.appSourceText(named: "WorkspaceCommandPaletteSurface.swift")
+        let rankerText = try Self.appSourceText(named: "WorkspaceCommandPaletteRanker.swift")
 
-        XCTAssertTrue(paletteText.contains("public struct WorkspaceCommandSurface"), "Command surface records should live beside palette ranking.")
+        XCTAssertTrue(paletteText.contains("public struct WorkspaceCommandSurface"), "Command surface records should live beside command palette API types.")
         XCTAssertTrue(paletteText.contains("public enum TopBarOverflowCommandCatalog"), "Top-bar overflow command projection should live beside command surfaces.")
-        XCTAssertTrue(paletteText.contains("public enum WorkspaceCommandPalette"), "Palette grouping and ranking should live in a focused command surface file.")
-        XCTAssertTrue(paletteText.contains("private static func score"), "Palette scoring should be directly guarded outside the aggregate surface file.")
+        XCTAssertTrue(paletteText.contains("public enum WorkspaceCommandPalette"), "Command palette API should stay in the focused command surface file.")
+        XCTAssertTrue(paletteText.contains("WorkspaceCommandPaletteRanker.rankedCommands"), "Public palette ranking should delegate to the focused ranker.")
+        XCTAssertTrue(paletteText.contains("WorkspaceCommandPaletteRanker.groupedCommands"), "Public palette grouping should delegate to the focused ranker.")
+        XCTAssertTrue(rankerText.contains("enum WorkspaceCommandPaletteRanker"), "Palette ranking/search should live in its own focused helper.")
+        XCTAssertTrue(rankerText.contains("private static func score"), "Palette scoring should be directly guarded in the ranker.")
+        XCTAssertTrue(rankerText.contains("private struct QueryRequest"), "Palette query scoping should stay with the ranker.")
+        XCTAssertFalse(paletteText.contains("private static func score"), "Command surface API should not own palette scoring internals.")
+        XCTAssertFalse(paletteText.contains("private struct QueryRequest"), "Command surface API should not own query scoping internals.")
         XCTAssertFalse(surfaceText.contains("public struct WorkspaceCommandSurface"), "WorkspaceSurface should not own command surface records.")
         XCTAssertFalse(surfaceText.contains("public enum TopBarOverflowCommandCatalog"), "WorkspaceSurface should not own top-bar overflow projection.")
         XCTAssertFalse(surfaceText.contains("public enum WorkspaceCommandPalette"), "WorkspaceSurface should not own command palette ranking.")
